+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ Next Release (TBD)
 
 * Alway overwrite existing API Gateway Rest API on updates
   (`#305 <https://github.com/awslabs/chalice/issues/305>`__)
+* Fix content type validation when charset is provided
+  (`#306 <https://github.com/awslabs/chalice/issues/306>`__)
 
 
 0.8.0